Adding alerts to add and deletes

+++ server/public/scripts/services/time-tracker.service.js
@@ -7,14 +7,26 @@ TimeTrackerApp.service('TimeTrackerService', ['$http', function($http){
     // POST request for sending entry data to database
     self.addEntry = function(newEntry){
         console.log('addEntry clicked', newEntry);
-        $http.post('/entries', newEntry)
-        .then(function(result){
+        swal({
+            title: "Add Task",
+            text: "Task Entered",
+            icon: "success",
+            buttons: true,
+          })
+          .then((addEntry) => {
+            if (addEntry) {
+            $http.post('/entries', newEntry)
+            .then(function(result){
             console.log('addEntry POST success', result);
             self.getEntries();
             self.getProjects();
         }).catch(function(error){
             console.log('addEntry POST error', error);
         }) // end $http POST route
+            } else {
+              swal("Oh no! Something went wrong!");
+            }
+          });
     } // end addEntry
 
     self.listOfEntries = {list: []};
@@ -34,13 +46,29 @@ TimeTrackerApp.service('TimeTrackerService', ['$http', function($http){
     // DELETE request to delete project from database
     self.removeEntry = function(entryId){
         console.log('In removeEntry');
-        $http.delete(`/entries/${entryId}`).then(function(result){
-            console.log('removeEntry DELETE success', result);
-            self.getEntries();
-            self.getProjects();
-        }).catch(function(error){
-            console.log('removeEntry DELETE ERROR', error);
-        }); // end $http DELETE route
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this task!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((removeEntry) => {
+            if (removeEntry) {
+              swal("Task Deleted", {
+                icon: "success",
+              });
+              $http.delete(`/entries/${entryId}`).then(function(result){
+                console.log('removeEntry DELETE success', result);
+                self.getEntries();
+                self.getProjects();
+            }).catch(function(error){
+                console.log('removeEntry DELETE ERROR', error);
+            }); // end $http DELETE route
+            } else {
+              swal("Task wasn't deleted");
+            }
+          });
     } // end removeEntry
 
 
@@ -48,14 +76,27 @@ TimeTrackerApp.service('TimeTrackerService', ['$http', function($http){
 
     // POST request for send project data to database
     self.addProject = function(newProject){
-        console.log('In addProject with new project', newProject);
-        $http.post('/projects', newProject).then(function(result){
-            console.log('addProject POST success', result);
-            self.getProjects();
-        }).catch(function(error){
-            console.log('addProject POST ERROR', error);
-        }); // end $http POST route
-    } // end addProjects
+        swal({
+            title: "Add Project",
+            text: "Project Entered",
+            icon: "success",
+            buttons: true,
+        })
+        .then((addProject) => {
+            if (addProject) {
+            console.log('In addProject with new project', newProject);
+            $http.post('/projects', newProject)
+            .then(function(result){
+                console.log('addProject POST success', result);
+                self.getProjects();
+            }).catch(function(error){
+                    console.log('addProject POST ERROR', error);
+            }); // end $http POST route
+            } else {
+                swal("Oh no! Something went wrong!");
+            } // end if addProject
+        });  
+    }  
 
     self.projectList = {list: []};
 
@@ -73,14 +114,29 @@ TimeTrackerApp.service('TimeTrackerService', ['$http', function($http){
 
     // DELETE request to delete project from database
     self.removeProject = function(projectId){
-        console.log('In removeProject');
-        $http.delete(`/projects/${projectId}`).then(function(result){
-            console.log('removeProject DELETE success', result);
-            self.getProjects();
-        }).catch(function(error){
-            console.log('removeProject DELETE ERROR', error);
-        })
-    }
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this project!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          }) // end sweet dialog
+          .then((removeProject) => {
+            if (removeProject) {
+              swal("Project Deleted", {
+                icon: "success",
+              }); // end sweet success
+              $http.delete(`/projects/${projectId}`).then(function(result){
+                console.log('removeProject DELETE success', result);
+                self.getProjects();
+            }).catch(function(error){
+                console.log('removeProject DELETE ERROR', error);
+            })
+            } else {
+              swal("Project wasn't deleted");
+            } // end if removeProject
+        })// end sweet delete
+    } // end removeProjects
 
     self.getProjects();
 }]);
